% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyph.support-de.R
\name{hyph.support.de}
\alias{hyph.support.de}
\title{Language support for German}
\usage{
hyph.support.de()
}
\description{
This function adds support for German to the sylly package. You should not
need to call it manually, as that is done automatically when this package is
loaded.
}
\details{
In particular,
      a new set of hyphenation patterns is being added (see \code{\link{hyph.de}}).
To use the patterns with \code{\link[sylly:hyphen]{hyphen}}, simply use the abbreviation:
\itemize{
  \item {\code{"de"}} {--- German hyphenation patterns (new spelling, since 1996)}
  \item {\code{"de.old"}} {--- German hyphenation patterns (old spelling, 1901--1996)}
}
}
\seealso{
\code{\link[sylly:hyphen]{hyphen}},
   \code{\link[sylly:set.hyph.support]{set.hyph.support}}
}
